add functionality to reorder input group identifiers based on the reference .fasta file
