+++ include/read_bitfield.hpp
@@ -12,6 +12,7 @@
 #include "KallistoFiles.hpp"
 
 void ReadClusterIndicators(std::istream &indicators_file, Reference &reference);
+void MatchClusterIndicators(const char delim, std::istream &groups, std::istream &fasta, Reference &reference);
 void ReadBitfield(KallistoFiles &kallisto_files, unsigned n_refs, std::vector<std::unique_ptr<Sample>> &batch, Reference &reference, bool bootstrap_mode);
 void ReadBitfield(const std::string &tinfile1, const std::string &tinfile2, const std::string &themisto_mode, const bool bootstrap_mode, const unsigned n_refs, std::vector<std::unique_ptr<Sample>> &batch);
 void VerifyGrouping(const unsigned n_refs, std::istream &run_info);
