+++ src/read_bitfield.cpp
@@ -7,6 +7,8 @@
 #include "bxzstr.hpp"
 #include "file.hpp"
 
+#include "tools/matchfasta.hpp"
+
 void VerifyGrouping(const unsigned n_refs, std::istream &run_info) {
   // Get the number of reference sequences in the pseudoalignment
   // contained in the 'n_targets' variable in run_info.json file.
@@ -77,6 +79,31 @@ void ReadClusterIndicators(std::istream &indicator_file, Reference &reference) {
     }
   } else {
     throw std::runtime_error("Could not read cluster indicators.");
+  }
+
+  reference.n_refs = reference.grouping.indicators.size();
+  reference.grouping.n_groups = str_to_int.size();
+}
+
+void MatchClusterIndicators(const char delim, std::istream &groups, std::istream &fasta, Reference &reference) {
+  std::unordered_map<std::string, unsigned> str_to_int;
+  std::vector<std::string> groups_in_fasta;
+  try {
+    mSWEEP::tools::matchfasta(groups, fasta, delim, &groups_in_fasta);
+  } catch (std::exception &e) {
+    throw std::runtime_error("Matching the group indicators to the fasta file failed, is the --groups-list delimiter correct?");
+  }
+
+  unsigned indicator_i = 0;
+  for (uint32_t i = 0; i < groups_in_fasta.size(); ++i) {
+    if (str_to_int.find(groups_in_fasta[i]) == str_to_int.end()) {
+      str_to_int[groups_in_fasta[i]] = indicator_i;
+      reference.group_names.emplace_back(groups_in_fasta[i]);
+      reference.grouping.sizes.emplace_back(0);
+      ++indicator_i;
+    }
+    ++reference.grouping.sizes[str_to_int[groups_in_fasta[i]]];
+    reference.grouping.indicators.emplace_back(str_to_int[groups_in_fasta[i]]);
   }
 
   reference.n_refs = reference.grouping.indicators.size();
